+++ package.json
@@ -30,11 +30,13 @@
     "react-dom": "^18.2.0",
     "react-responsive": "^9.0.2",
     "sharp": "^0.31.3",
-    "tailwindcss": "3.2.7"
+    "tailwindcss": "3.2.7",
+    "zlib-sync": "^0.1.8"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
     "eslint-config-next": "^13.4.19",
+    "node-loader": "^2.0.0",
     "postcss": "^8.4.21",
     "prettier": "3.0.1",
     "tailwindcss": "^3.2.7"
